// src/servers/customers/types/Address.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Address = {
  /**
   * @description Unique identifier for the address
   * @type string | undefined, uuid
   */
  addressId?: string
  /**
   * @description Full address
   * @type string
   */
  address: string
  /**
   * @description Postal code in the format 1234-567
   * @type string
   */
  postalCode: string
  /**
   * @description City name
   * @type string
   */
  city: string
  /**
   * @description Phone number in international format
   * @type string
   */
  phoneNumber: string
  /**
   * @description Taxpayer identification number (9 digits)
   * @type string | undefined
   */
  nif?: string
  /**
   * @description Indicates if the taxpayer is added
   * @type boolean | undefined
   */
  addTaxpayer?: boolean
  /**
   * @description Indicates if this is the default address
   * @type boolean | undefined
   */
  defaultAdress?: boolean
  /**
   * @description Indicates if this is the default billing address
   * @type boolean | undefined
   */
  defaultAdressFaturation?: boolean
  /**
   * @description Unique identifier for the customer
   * @type string | undefined, uuid
   */
  customerId?: string
}